Adapted genetic algorithm to do simulation-based evolution (vs. supervised learning)
- Created a generic Simulation class that takes a gym-complient simulation environment
as input and uses the network to guide the actions. Only available mode for now is one
output neuron per possible action and chosing the argmax action.
- Modified the algo so that calling run() multiple times keeps evolving the same
population.
- Improved logging and aligned it with what is done in SimpleRL: accepts a "wandb" arg
to log directly to W&B, counts number of episodes (i.e total simulations run/supervised
examples trained on) in addition to the number of generations (which allows a fair
comparison with traditional Q-learning-like learning algorithms), logs the score
distribution among the population and across the multiple simulations for the best
individual.