Use Simulation class in qlearning's basesimulation
- Standardized findAction() method in the structures, and moved there the
part of the action selection logic that was in the simulations.
- The QLearning (and similar algos)'s learnStep() stage is now called from
the simulation on every step, as an optional callback parameter that is
passed to the performEpisode() method.
- Made all evaluation functions (simulation & loss) inherit from the base
EvalFunc class and implement isLoss/isSimulation property.
- Renamed Simulation's apply() method to performEpisodes() to be more clear,
and made the nEpisodes parameter optional (can now be passed to constructor).
- Integrated the supervised learning examples (xs & ys) into the loss function
and adapted the code elsewhere to reflect this change.
- Moved the evaluation functions out of the neural network itself and passing
them directly to the optimization algorithms.
- Adapted structures' copy() methods to take into account the new attributes
(actionSelectionMode & isLearning).
- Potentially other minor changes and bugfixes.